fixed a bug related to the embedding matrix dtype. Has to be float32

+++ pytorch_widedeep/utils/text_utils.py
@@ -138,4 +138,4 @@ def build_embeddings_matrix(
             )
         )
 
-    return embedding_matrix
+    return embedding_matrix.astype("float32")
